Delay resolution of peers' IP addresses.
In environments where the IPs a hostname resolves to can change (e.g.,
Elastic IPs in AWS), resolving IP addresses once at startup means that
if the mapping changes, all nodes in the ensemble must be restarted. To
handle this case, IPs are only resolved when attempting to connect to
peers.